+++ src/metatrain/experimental/nativepet/tests/test_regression.py
@@ -56,90 +56,15 @@ def test_regression_init():
     )
 
     # if you need to change the hardcoded values:
-    torch.set_printoptions(precision=12)
-    print(output["mtt::U0"].block().values)
-
-    torch.testing.assert_close(output["mtt::U0"].block().values, expected_output)
-
-
-def test_regression_energies_train():
-    """Perform a regression test on the model when
-    trained for 2 epoch on a small dataset, energies only"""
-
-    # reproducibility
-    random.seed(0)
-    np.random.seed(0)
-    torch.manual_seed(0)
-
-    systems = read_systems(DATASET_PATH)
-
-    conf = {
-        "mtt::U0": {
-            "quantity": "energy",
-            "read_from": DATASET_PATH,
-            "reader": "ase",
-            "key": "U0",
-            "unit": "eV",
-            "type": "scalar",
-            "per_atom": False,
-            "num_subtargets": 1,
-            "forces": False,
-            "stress": False,
-            "virial": False,
-        }
-    }
-    targets, target_info_dict = read_targets(OmegaConf.create(conf))
-    dataset = Dataset.from_dict({"system": systems, "mtt::U0": targets["mtt::U0"]})
-
-    hypers = DEFAULT_HYPERS.copy()
-    hypers["training"]["num_epochs"] = 2
-    hypers["training"]["scheduler_patience"] = 1
-
-    dataset_info = DatasetInfo(
-        length_unit="Angstrom", atomic_types=[1, 6, 7, 8], targets=target_info_dict
-    )
-    model = NativePET(MODEL_HYPERS, dataset_info)
-
-    trainer = Trainer(hypers["training"])
-    trainer.train(
-        model=model,
-        dtype=torch.float32,
-        devices=[torch.device("cpu")],
-        train_datasets=[dataset],
-        val_datasets=[dataset],
-        checkpoint_dir=".",
-    )
-
-    # Predict on the first five systems
-    systems = [system.to(torch.float32) for system in systems]
-    for system in systems:
-        get_system_with_neighbor_lists(system, model.requested_neighbor_lists())
-    output = model(
-        systems[:5],
-        {"mtt::U0": ModelOutput(quantity="energy", unit="", per_atom=False)},
-    )
-
-    expected_output = torch.tensor(
-        [
-            [3.012795209885],
-            [1.843898773193],
-            [0.775631427765],
-            [1.857149720192],
-            [1.013873577118],
-        ]
-    )
-
-    # if you need to change the hardcoded values:
-    torch.set_printoptions(precision=12)
-    print(output["mtt::U0"].block().values)
+    # torch.set_printoptions(precision=12)
+    # print(output["mtt::U0"].block().values)
 
     torch.testing.assert_close(output["mtt::U0"].block().values, expected_output)
 
 
 def test_regression_energies_forces_train():
-    """Perform a regression test on the model when
-    trained for 2 epoch on a small dataset with energies
-    and forces"""
+    """Perform a regression test on the model when trained for 2 epoch on a small "
+    "dataset with energies and forces"""
 
     # reproducibility
     random.seed(0)
@@ -210,9 +135,9 @@ def test_regression_energies_forces_train():
     )
 
     # if you need to change the hardcoded values:
-    torch.set_printoptions(precision=12)
-    print(output["energy"].block().values)
-    print(output["energy"].block().gradient("positions").values.squeeze(-1)[0])
+    # torch.set_printoptions(precision=12)
+    # print(output["energy"].block().values)
+    # print(output["energy"].block().gradient("positions").values.squeeze(-1)[0])
 
     torch.testing.assert_close(output["energy"].block().values, expected_output)
     torch.testing.assert_close(