+++ docs/src/dev-docs/changelog.rst
@@ -32,7 +32,7 @@ Fixed
 Added
 #####
 
-- ``NanoPET`` and ``NativePET`` can now train on symmetric rank-2 Cartesian tensors
+- ``NanoPET`` and ``NativePET`` can now train on non-conservative stresses
 
 Version 2025.5 - 2025-04-13
 ---------------------------